+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- Add `OpenProject/NoSleepInFeatureSpecs` cop to check that `sleep` calls in
+  feature specs are not greater than 1 second.
+
 ## [0.1.0] - 2024-07-05
 
 - Initial release